PID v1.0

+++ STM32Cube/Components/Inc/hcsr04_sensor.h
@@ -33,7 +33,7 @@ struct us_sensor_str
 
 	TIM_Channel trig_channel;
 
-	volatile uint32_t distance_cm;
+	volatile float distance_cm;
 };
 
 
@@ -53,9 +53,9 @@ void hc_sr04_init(struct us_sensor_str *us_sensor, TIM_HandleTypeDef *htim_echo,
 *@param:
 *@retval: None
 */
-uint32_t hc_sr04_convert_us_to_cm(uint32_t distance_us);
+float hc_sr04_convert_us_to_cm(float distance_us);
 
-int position(float dis1, float dis2, float pos);
+float position(float dis1, float dis2, float pos);
 
 #ifdef __cplusplus
 }
